requirements.txt
numpy==1.26.4
pandas==2.2.0
pillow==10.2.0
matplotlib==3.8.2
opencv-python==4.9.0.80
